New feature: Delayed timers
This makes statsd able to collect timer stats that occurred a specified
time in the past, instead of only in the last 10 seconds.  By default,
stats can come in from up to 5 minutes in the past. The statsd client
needs to be updated to send a timestamp along with the timer value,
details are in the readme